Advanced version to 0.002.
I am planning on using even/odd versions for stable and unstable versions,
respectively.  When sufficient features are finished and bugs are squashed,
the version will be incremented to an even number and released.  Immediately
after, the version will be incremented again, to an odd value, and development
work resumes.  This is particularly to have stuff done by the script show as a
different version from the previous during development.