Fetch the RefMatch user and store the id.
We want this script's database modifications to be associated with a user, so
fetch the RefMatch user's information and store its id.  This is not the best
(or perhaps even good) way to get the user's id.  It at least should not be
fetching all of the columns from the user's table, including the password
hash.