Get the number of incomplete references in one step.
We will ultimately be running a more complex query to get actual work
references from the database, so run a quick query to get the number of
references which do not have a referenced_work_id.  Do not store this in a
result set, as it will not be used for anything other than getting the count.