Get the result set of pending items, with metadata.
Get all of the work references which do not have a referenced_work_id, but
also get references that have a guess from this algorithm.  Include metadata
from the referenced_work_guesses table, and sort by the last time the
reference was checked.

Of note here, is that the sort order specifies null values for last_checked
should come first (not the default behavior).  This will allow us to
prioritize references which have never been checked, then re-check the oldest
guesses.

Also note that this query does not limit itself by the user who made the check
with the same algorithm.  This might change in the future.