Create the current timestamp in Postgres format.
I was having trouble setting up an update query which uses Postgres' now()
function to get the current time on the server.  This is the next best thing,
then, as it gets the current timestamp, and converts it into a format that
Postgres can understand.  It also uses a time zone, and here it is set to a
reasonable value for me.  The default time zone is UTC, and this would almost
work as well for me, as the actual "time" is the same, but originally inserted
as from a different time zone.