Update or create the guess for the reference.
This query first runs a select statement to see if the row exists (honoring
the unique constraint on the columns).  If none exist, it then immediately
inserts a new record.  If one does, that record is updated.

There is an alternate method, update_or_new, which at least does not run the
insert or update query immediately.  This version can also be used to
determine, from the script, if the record had already existed.