Tell the DB connection to expect UTF-8.
Added a setting to the connect method call to enable UTF-8 in the connection
with the PostgreSQL database.  Without this, UTF-8-encoded text was being
interpreted as ASCII (presumably).  This was even with the -CSD command line
switch (see http://perldoc.perl.org/perlrun.html#-C for information on the -C
switch), which made Perl assume input and output for standard streams and
other (file) streams were in UTF-8.  This did not affect the database
connection, probably, since that connection is not a standard stream or file
stream.

See http://wiki.catalystframework.org/wiki/tutorialsandhowtos/Using_Unicode.inline
for more information on using UTF-8 in a Catalyst application.  Other types of
databases use different settings.  For example, use "mysql_enable_utf8 => 1"
for MySQL, and "unicode => 1" for SQLite.