Check if citation does not have a title.
If ParsCit cannot get a title from a reference, the getTitle accessor method
returns null.  Check that this value is true (is defined and is not an empty
string) before checking its length.  If either test indicate we do not have a
title, use the full reference text in comparisons.