Use a dummy citation if one was not extracted.
For unusual citations, such as a web URL, ParsCit might not be able to extract
a citation.  When this happens, the array of citations is empty (when only one
possible citation was in the file to parse).  Create a "dummy" citation object
directly, and set its string and title to be the full reference text.